+++ R/fscores.R
@@ -17,7 +17,9 @@
 #' 
 #' If the input object is a discrete latent class object estimated from \code{\link{mdirt}}
 #' then the returned results will be with respect to the posterior classification for each 
-#' individual. 
+#' individual. The method inputs for \code{'DiscreteClass'} objects may only be \code{'EAP'},
+#' for posterior classification of each response pattern, or \code{'EAPsum'} for posterior
+#' classification based on the raw sum-score.
 #'
 #'
 #' @aliases fscores